requirements.txt
discord.py-self>=1.9.0
python-dotenv>=1.0.0

